requirements.txt
# Core dependencies
streamlit>=1.31.0
langchain>=0.1.0
langchain-community>=0.0.21 # Added explicitly
langchain-openai>=0.0.2
openai>=1.10.0
python-dotenv>=1.0.0
httpx>=0.24.1
tiktoken>=0.4.0
pydantic>=2.0.0
rich>=13.4.2
tenacity>=8.2.2
pandas>=1.3.0 # Added for DataFrame use in ontox_client

# Optional testing dependencies
# fastapi>=0.100.0
# uvicorn>=0.22.0
